Move documentation of the options file into the file.
Most notably, allow the options file (and the processing) to specify
multiple groups of files, each containing lists of file selectors.  One
we start applying the phenotypes, we will want to compare specific users
- tongue rollers vs. non-tongue-rollers, for example.  So the file
format now allows us to specify a first group containing only those with
the "tongue roller" phenotype and a second group containing only those
who are classed as non-tongue rollers.

Added optional output of the filenames of selected files and files with
no data to process that meets the selection parameters.  Modified the
option to print line number detail so that we can limit the output to
every nth line.

Add constants for the selection parameter names.

The internal parsed structure is now
{rsids:{phonotype_label:{(chromosome_number,position):{genotype:count}}}}