Created a prototype of phenotype parser. It parses phenotypes.csv in the same folder (hardcoded file name). The output is a JSON file with the following structure: an array of value objects, containing phenotype, phenotype_id,data. Data is an array of user_id and a response to phenotype questionnaire.
The results are not sanitized yet.

Expected use: A json file is loaded, a phenotype is found (typically by id), data is extracted for that phenotype. You must iterate over data and check response values. If you found a value that you are interested in, get the "user_id" from the same dictionary and use it to fetch a snp file.